Moved dataset related classes to a separate file
